pyproject.toml
[project]
name = "lang-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.31",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langsmith>=0.4.16",
    "mcp>=1.13.1",
    "python-dotenv>=1.1.1",
]
